Use a fork of spray-json for now. See below.
Scala's case classes (which is what we parse JSON responses into)
support up to 22 parameters as of this commit. However, spray-json
hasn't yet been updated to support 22 parameters. It only supports
15. spray-json2 will support up to 22, but isn't released yet.

So I've taken someone's fork (who added params 16-22), published it
to a maven repo, then uploaded it to fedorapeople.

The fork we're using is here: https://github.com/rditerwich/spray-json

This is only temporary, until either upstream releases a spray-json
with support for 22 params, or until spray-json2 comes out and we can
update to it.